MultiChannelGroupByHashInlineFastBBAllTypes
